requirements.txt
# Fundamental
# numpy
# jax  # See document for installation
optax
dm-haiku
numpyro
numba
tensorboardX
ipdb

# Used by environment
gymnasium[mujoco]==1.0.0
dill==0.3.8
# free-mujoco-py
matplotlib
scikit-learn

# Utility
tqdm
setproctitle
